Fixed optimized d*-lite algorithm
Issue was that we would record the first change to a node in the map but not subsequent changes. Thus when we saw a new node be created as uncertain, we would record uncertain, but then if that node was confirmed an obstacle before asking for the node diff, it would ignore that we know the node changed to an obstacle. This is why we saw behavior such as planning to go through a wall, but the agent getting stuck as the fringe was not expanded and the min g value was the last node.